move schema to other file
