+++ libs/language-server/capabilities/code-completion/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/get-completion-items';
+export * from './lib/schema-completion';
